Simplify folder structure layout
