+++ __mocks__/aws-sdk.js
